Update: requestCollectionsAPI path changed from App.js to Main.js
